+++ tests/test_fixtures.py
@@ -1,8 +1,7 @@
 """Test fixtures."""
 
 import vehiclepass
-
-from .conftest import mock_responses
+from tests.utils import mock_responses
 
 
 @mock_responses(